[wps] add storage functions
1. exec_show_storage_imformation
2. exec_show_store_list
3. exec_find_item_in_storage

+++ supermarket.sql
@@ -205,7 +205,7 @@ create table storage
 /*==============================================================*/
 create table storelist
 (
-   item_id              bigint not null,
+   item_id              int not null,
    stor_id              int not null,
    number               int not null,
    primary key (item_id, stor_id)
